+++ Calculator/Operator.cpp
@@ -1,6 +1,6 @@
 #include "Operator.h"
 
-COperator::COperator(string _strOrigin, int _iPriority, function<int(int,int)> _pFunction) :
+COperator::COperator(string _strOrigin, int _iPriority, function<double(double, double)> _pFunction) :
     CToken(_strOrigin, _iPriority), m_pFunction(_pFunction)
 {
 }
